+++ CHANGELOG.md
@@ -6,5 +6,6 @@
 - FIX: Reset WX Data 
 - FIX: missing APRS data if using Radio and not IGate.
 - ADD: Set IGate filters under Settings.
+- ADD: Multi Map Provider support.